requirements/cms.txt
coverage==4.0.1
Django==1.8.5
django-bootstrap3==6.2.2
django-classy-tags==0.7.1
django-cms==3.2.1
django-contact-form==1.1
django-formtools==1.0
django-mptt==0.8.0
django-sekizai==0.9.0
django-treebeard==4.0
djangocms-admin-style==1.1.0
djangocms-file==0.2.0
djangocms-picture==0.2.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.8.1
html5lib==0.9999999
Pillow==3.1.1
six==1.10.0
wheel==0.24.0
